♻️(models) reorganize models inside a directory
This avoids having an akward "base_models.py" file
at the root of the app. We take this opportunity to
split models in 2 groups: account and video.
