+++ marsha/core/models/__init__.py
@@ -1,0 +1,4 @@
+"""Make all models available from marsha.core.models."""
+# pylint: disable=wildcard-import,unused-wildcard-import
+from .account import *  # noqa isort:skip
+from .video import *  # noqa isort:skip
